# backend/hhrr_app/static/swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "This API provides the Plena Global Holdings software development team with the necessary knowledge base for the integration of the different applications with the services offered by the human resources department.",
		"version": "1.0.0",
		"title": "HUMMAN RESOURCES DEPARTMENT API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "/api/v1/"
		}
	],
	"components": {
		"schemas": {
...
			"tbl_role": {
				"type": "object",
				"properties": {
					"ccn_role": {
						"type": "integer (pk) auto_increment",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"area": {
						"type": "string",
						"minimum": 0,
						"maximum": 50,
						"example": "HUMAN RESOURCES",
						"description": "String Field"
					},
					"role": {
						"type": "string",
						"minimum": 0,
						"maximum": 50,
						"example": "GENERAL SERVICES WORKER",
						"description": "String Field"
					},
					"process": {
						"type": "string",
						"minimum": 0,
						"maximum": 50,
						"example": "HUMAN RESOURCES",
						"description": "String Field"
					},
					"full_role": {
						"type": "string",
						"minimum": 0,
						"maximum": 250,
						"example": "HUMAN RESOURCES - HUMAN RESOURCES",
						"description": "String Field"
					}
				}
			},
...
			"tbl_schooling_level": {
				"type": "object",
				"properties": {
					"ccn_schooling_level": {
						"type": "integer (pk) auto_increment",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"description_schooling_level": {
						"type": "integer",
						"minimum": 0,
						"maximum": 40,
						"example": 5001,
						"description": "DIAN identification city name"
					}
				}
			},
			"tbl_race": {
				"type": "object",
				"properties": {
					"ccn_race": {
						"type": "integer (pk) auto_increment",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"description_race": {
						"type": "string",
						"minimum": 0,
						"maximum": 40,
						"example": "POSTGRADO",
						"description": "Employee schooling level"
					}
				}
			},
...
		}
	},
	"paths": {
...
		"/role": {
			"post": {
				"tags": [
					"Basic data from role"
				],
				"description": "Create a new role to human resource platform",
				"requestBody": {
					"description": "Post New role to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_role"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from role"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					}
				}
			}
		},
		"/role/{ccn_role}": {
			"get": {
				"tags": [
					"Basic data from role"
				],
				"parameters": [
					{
						"name": "ccn_role",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from role"
				],
				"parameters": [
					{
						"name": "ccn_role",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_role"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from role"
				],
				"parameters": [
					{
						"name": "ccn_role",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_role"
						}
					}
				}
			}
		},
...
		"/race": {
			"post": {
				"tags": [
					"Basic data from Race"
				],
				"description": "Create a race data to human resource platform",
				"requestBody": {
					"description": "Post race data to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_race"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_race"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_race"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from Race"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
            "$ref": "#/components/schemas/tbl_race"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_race"
						}
					}
				}
			}
		},
		"/race/{ccn_race}": {
			"get": {
				"tags": [
					"Basic data from Race"
				],
				"parameters": [
					{
						"name": "ccn_race",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_race"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
						"$ref": "#/components/schemas/tbl_race"
						}
					}
				}
			},
			"put": {
				"tags": [
				"Basic data from Race"
				],
				"parameters": [
					{
						"name": "ccn_race",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
              "$ref": "#/components/schemas/tbl_race"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
            "$ref": "#/components/schemas/tbl_race"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
            "$ref": "#/components/schemas/tbl_race"
						}
					}
				}
			},
			"delete": {
				"tags": [
        "Basic data from Race"
				],
				"parameters": [
					{
						"name": "ccn_race",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
            "$ref": "#/components/schemas/tbl_race"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
            "$ref": "#/components/schemas/tbl_race"
						}
					}
				}
			}
		},
		"/schooling_level": {
			"post": {
				"tags": [
					"Basic data from SchoolingLevel"
				],
				"description": "Create a schoolingLevel data to human resource platform",
				"requestBody": {
					"description": "Post schoolingLevel data to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_schooling_level"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from SchoolingLevel"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					}
				}
			}
		},
		"/schooling_level/{ccn_schooling_level}": {
			"get": {
				"tags": [
					"Basic data from SchoolingLevel"
				],
				"parameters": [
					{
						"name": "ccn_schooling_level",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from SchoolingLevel"
				],
				"parameters": [
					{
						"name": "ccn_schooling_level",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_schooling_level"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from SchoolingLevel"
				],
				"parameters": [
					{
						"name": "ccn_schooling_level",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_schooling_level"
						}
					}
				}
			}
		}
	}
}